+++ urbanstats/geometry/shapefiles/shapefiles/cities.py
@@ -1,7 +1,7 @@
 import us
 
 from urbanstats.geometry.shapefiles.shapefile import Shapefile
-from urbanstats.universe.universe_provider.contained_within import us_domestic_provider
+from urbanstats.universe.universe_provider import us_domestic_provider
 
 CITIES = Shapefile(
     hash_key="census_places_6",
