+++ urbanstats/geometry/shapefiles/shapefiles/judicial.py
@@ -2,7 +2,7 @@ import geopandas as gpd
 import us
 
 from urbanstats.geometry.shapefiles.shapefile import Shapefile
-from urbanstats.universe.universe_provider.contained_within import us_domestic_provider
+from urbanstats.universe.universe_provider import us_domestic_provider
 
 
 def render_ordinal(x):
