+++ urbanstats/geometry/shapefiles/shapefiles/native.py
@@ -2,7 +2,7 @@ import geopandas as gpd
 from permacache import permacache
 
 from urbanstats.geometry.shapefiles.shapefile import Shapefile
-from urbanstats.universe.universe_provider.contained_within import us_domestic_provider
+from urbanstats.universe.universe_provider import us_domestic_provider
 
 
 def is_native_statistical_area(row):
