+++ urbanstats/geometry/shapefiles/shapefiles/neighborhoods.py
@@ -1,7 +1,7 @@
 import us
 
 from urbanstats.geometry.shapefiles.shapefile import Shapefile
-from urbanstats.universe.universe_provider.contained_within import us_domestic_provider
+from urbanstats.universe.universe_provider import us_domestic_provider
 
 NEIGHBORHOODS = Shapefile(
     hash_key="zillow_neighborhoods_6",
