+++ urbanstats/geometry/shapefiles/shapefiles/school_districts.py
@@ -6,7 +6,7 @@ from permacache import permacache
 
 from urbanstats.geometry.shapefiles.shapefile import Shapefile
 from urbanstats.geometry.shapefiles.shapefiles.counties import COUNTIES
-from urbanstats.universe.universe_provider.contained_within import us_domestic_provider
+from urbanstats.universe.universe_provider import us_domestic_provider
 
 
 @permacache("population_density/shapefiles/school_district_shapefiles")
