+++ urbanstats/geometry/shapefiles/shapefiles/urban_centers.py
@@ -6,8 +6,10 @@ import us
 from urbanstats.geometry.shapefiles.shapefile import Shapefile
 from urbanstats.special_cases.ghsl_urban_center import load_ghsl_urban_center
 from urbanstats.universe.universe_list import get_universe_name_for_state
-from urbanstats.universe.universe_provider import INTERNATIONAL_PROVIDER
-from urbanstats.universe.universe_provider.contained_within import us_domestic_provider
+from urbanstats.universe.universe_provider import (
+    INTERNATIONAL_PROVIDER,
+    us_domestic_provider,
+)
 from urbanstats.universe.universe_provider.universe_provider import UniverseProvider
