+++ urbanstats/geometry/shapefiles/shapefiles/usda_county_type.py
@@ -3,7 +3,7 @@ import pandas as pd
 from permacache import permacache
 
 from urbanstats.geometry.shapefiles.shapefile import Shapefile
-from urbanstats.universe.universe_provider.contained_within import us_domestic_provider
+from urbanstats.universe.universe_provider import us_domestic_provider
 
 
 @permacache("population_density/shapefiles/usda_county_type")
