+++ urbanstats/universe/universe_provider/__init__.py
@@ -8,9 +8,10 @@ from urbanstats.universe.universe_provider.constant_provider import (
     ConstantUniverseProvider,
 )
 from urbanstats.universe.universe_provider.contained_within import (
+    STATE_PROVIDER,
     ContainedWithinUniverseProvider,
 )
-
+from urbanstats.universe.universe_provider.override import OverrideUniverseProvider
 
 INTERNATIONAL_PROVIDER = CombinedUniverseProvider(
     [
@@ -19,3 +20,12 @@ INTERNATIONAL_PROVIDER = CombinedUniverseProvider(
         UrbanCenterStateUniverseProvider(),
     ]
 )
+
+
+def us_domestic_provider(overrides=None):
+    state_provider = STATE_PROVIDER
+    if overrides is not None:
+        state_provider = OverrideUniverseProvider(overrides, state_provider)
+    return CombinedUniverseProvider(
+        [ConstantUniverseProvider(["world", "North America", "USA"]), state_provider]
+    )
